major restructure WIP: tests not building properly
