+++ CMakeLists.txt
@@ -1,8 +1,5 @@
 cmake_minimum_required(VERSION 3.17)
 project(combat_cpp VERSION 0.1.0 LANGUAGES CXX)
-set(CMAKE_CXX_STANDARD 17)
-set(CMAKE_CXX_STANDARD_REQUIRED ON)
-
 
 #----------------------------------------------------------------
 # Packages
@@ -28,81 +25,28 @@ FetchContent_MakeAvailable(gtest)
 add_subdirectory(messages)
 
 #----------------------------------------------------------------
-# Main Sources
+# Create A Library
 #----------------------------------------------------------------
-set(ser_includes
-    include/Die.hpp
-    include/Combatant.hpp
-    include/Utility.hpp
-    include/Item.hpp
-    include/Weapon.hpp
-    include/Armour.hpp
-    include/States.hpp
-    include/GameLogic.hpp
+file(GLOB sources "${CMAKE_CURRENT_SOURCE_DIR}/src/*.cpp")
+
+add_library(${PROJECT_NAME} ${sources})
+
+target_include_directories(${PROJECT_NAME}  
+    PUBLIC 
+    ${CMAKE_CURRENT_SOURCE_DIR}/include
 )
 
-set(ser_sources
-    src/Die.cpp
-    src/Combatant.cpp
-    src/Utility.cpp
-    src/Weapon.cpp
-    src/Armour.cpp
-    src/GameLogic.cpp
-)
-
-set(all_sources
-    ${ser_includes}
-    ${ser_sources}
-)
-
-
-#----------------------------------------------------------------
-# Executables
-#----------------------------------------------------------------
-add_executable(exampleMain
-    src/main.cpp
-    ${all_sources}
-)
-
-target_compile_features(exampleMain
+target_compile_features(${PROJECT_NAME}
     PRIVATE
         cxx_std_17
 )
 
-target_include_directories(exampleMain
-    PUBLIC
-        ${CMAKE_CURRENT_SOURCE_DIR}/include
-    PRIVATE
-        ${CMAKE_CURRENT_SOURCE_DIR}/src
-)
+#----------------------------------------------------------------
+# Add Apps
+#----------------------------------------------------------------
+add_subdirectory(app)
 
-target_link_libraries(exampleMain
-    PUBLIC
-        messages
-)
-
-add_executable(runAllTests
-    src/all_tests.cpp
-    ${all_sources}
-)
-
-target_compile_features(runAllTests
-    PRIVATE
-        cxx_std_17
-)
-
-target_include_directories(runAllTests
-    PUBLIC
-        ${CMAKE_CURRENT_SOURCE_DIR}/include
-    PRIVATE
-        ${CMAKE_CURRENT_SOURCE_DIR}/src
-)
-
-target_link_libraries(runAllTests
-    PUBLIC
-        gtest_main
-        messages
-)
-
-enable_testing()
-add_test(GTest runAllTests)
+#----------------------------------------------------------------
+# Add Tests
+#----------------------------------------------------------------
+add_subdirectory(test)
