+++ include/combat_cpp/Die.hpp
