+++ include/combat_cpp/Utility.hpp
