+++ src/Utility.cpp
@@ -1,4 +1,4 @@
-#include "Utility.hpp"
+#include "combat_cpp/Utility.hpp"
 
 #include <algorithm>
 #include <numeric>
